requirements-minimal.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
python-dotenv>=1.0,<2.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
